requirements.txt
annotated-types==0.7.0
anyio==4.10.0
byteplus-python-sdk-v2==3.0.8
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Pillow==10.4.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.5
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
